feat: 增加了对修改用户address的校验

+++ model/rdm/req.go
@@ -16,7 +16,7 @@ type UpdateUserinfoReq struct {
 	Gender    string `json:"gender"`
 	AvatarURL string `json:"avatar_url"`
 	PhoneNum  string `json:"phone_num"`
-	AddressId int    `json:"address_id,string"`
+	AddressId int64  `json:"address_id,string"`
 }
 
 type GetAddressListReq struct {
